SCRD-1855 Address SLES a2ensite security concerns
The added a2ensite script is called with elevated privileges so we need
to ensure that is implemented in a security aware fashion. To this end
we need to enhance the script, and how it is installed, to address the
concerns identified.

Security issues addressed in this patch:
  * Use restricted mode for the shell
  * Explicitly set PATH to a safe value
  * Explicitly set IFS to a safe value
  * Quote instances of command line arguments
  * Quote all uses of all variables
  * Use long option names for ln arguments
  * Use full path when creating symbolic link target.
  * Install file with 0500 as mode setting

Also updated code to be consistent in use of ${...} for all variable
dereferences.

Change-Id: I533b556d895bfe2633fb1a04946479ffe0db080b