SCRD-2385 Setup OS-dependent Apache2 vhost dir
On SUSE operating systems it is more natural to place all enabled
virtual host config files in `/etc/apache2/vhosts.d`, as opposed to the
Debian-style configuration which places virtual host configs in a
staging area and then enables them with a utility called `a2ensite`
which creates a symlink between the staging area and the active config
directory. Up until now we've been forcing the Debian-style method on
SLES since historically that is what Ardana has been built for. This
patch starts the process of converting deployments on SLES to use its
native style of Apache configuration by setting a conditional fact to
denote the proper vhost config directory. Other playbooks will have to
replace hard-coded usage of `/etc/apache2/sites-available` with
`{{ apache2_vhost_dir }}` and stop using the `a2ensite` and `a2dissite`
utilities when operating on a SLES host.

Change-Id: I9c739dceb87055263b1d7903ebb5b8d617b4dcd3