SCRD-2848 Add aliased directories in site template
Add an optional variable, additional_dirs, that represents a dictionary
of filesystem directories and the Apache aliases that will link to them.
This will allow us to support hosting repository directories at
different locations on the filesystem using the same listening port. The
reason for keeping them separate rather than bundling all repositories
under the existing structure is that the Cloud and SLES media has the
potential to be manually configured by a customer prior to running
ardana-init, and we do not want this customer-managed directory or the
setup of the virtualenv or extras servers to be intertangled or
dependent on one another, but we do want to take advantage of the
existing site.

Change-Id: Iea38bbe377b1375bd40323eb14f996484b598973