Fix check for haproxy being active (bsc#1088433)
A killall -0 checks for *any* process that is named haproxy,
and neutron also starts haproxy processes, so neutron active
(but haproxy being down) would pass the check, which is
unintentional. we need to check for the system provided one
to run. Also fix some issues with the script security, specifying
commands without full path is logged as SECURITY VIOLATION.

Change-Id: Iadcdb0a6651167946e36a70117817516ee47311f