haproxy: configure nbproc on virtual setup (bsc#1094275)
On virtual deployments, the nbproc global setting is
not configured and defaults to a value of 1, but the
bind-process directive is still erroneously configured
to values greater than 1. This leads to warnings such
as the following:

[WARNING] 141/091003 (5826) : Proxy 'ardana-ccp-vip-FRE-API-mgmt-9090': the process specified on the 'bind-process' directive refers to a process number that is higher than global.nbproc. The proxy has been forced to run on process 1 only.

The present patch corrects this problem by configuring
nbproc on virtual deployments the same way as on bare metal
deployments.

Change-Id: I9214c84d7fb652ccb63364da7bd9d13cc34c96ff