Fix haproxy not being able to log using journald (bsc#1005886)
In order to address CVE-2016-8611 properly, haproxy should log an
entry when the glance-api rate limiter it triggeed.  In testing the
rate limiter, it was discovered haproxy is not logging anything currently,
other than its starup and shutdown messages. This patch addresses the
issue by adding another socket located insisde the haproxy chroot
jail. It also cleans up some deprecated options for contolling the
log enries only going to the desired log and not also syslog.

https://www.rsyslog.com/doc/v8-stable/compatibility/
v7compatibility.html#omruleset-and-discard-action-are-deprecated

Change-Id: I22a4c9bfd84f0711c61a21f6582272237cfc9793