Use bool filter to ensure valid boolean evaluation (SOC-11192)
Fixes: SOC-11192, SOC-11193

When investigating issues related to SOC-11169 I saw failures in the
FND-AP2 role's start.yml and status.yml task lists when attempting to
manage apache2 when there were no active vhosts, even though checking
exists to guard against doing so.

After reviewing the code I noticed that it was testing to see if a
fact was "True" using the "== True" expression. This is not a safe
construct as depending on how the "True" value is assigned to a fact;
if it ends being assigned as the string value "True" then you end up
comparting "True" to the boolean True, which aren't equal.

However there exists a bool filter to overcome this issue which fixes
the task lists to work correctly when no vhosts are enabled.

Similarly we should also always set the value of apache2_sitesenabled
based upon the latest results of running the apache2ctl command, to
avoid issues that can arrise if the relevant tasks have been included
in the running ansible context previously.

Finally since the affected code is duplicated in the start.yml and
status.yml tasks lists, it has been separated out into a separate
check_sites_enabled.yml task list, that is included in the other task
lists, to simplify maintence of the code.

Change-Id: Ie9e23b413f68eb23d356d3c056e90828c837f502